Included file get parsed once.
Create a new method: initialize_lines that receives the lines to
be parsed and make sure that lines variable is an array or if
it's a String split it with "\n".

That method is used when a line includes a file, instead of
create a new Orgmode::Parser and parse that file to get the lines
and parse them again.